Redesign circ API

+++ include/phase2/data.h
@@ -52,7 +52,7 @@ void data_close(data_id);
  *   0			if the value was successfully retrieved
  *  -1			in case of error
  */
-int data_multidet_getnums(data_id fid, size_t *num_qubits, size_t *num_dets);
+int data_multidet_getnums(data_id fid, uint32_t *num_qubits, size_t *num_dets);
 
 /**
  * Perform action "op" on each determinant in "multidet" group.
@@ -99,7 +99,7 @@ int data_multidet_foreach(data_id fid,
  *   0			if the value was successfully retrieved
  *  -1			in case of error
  */
-int data_hamil_getnums(data_id fid, size_t *num_qubits, size_t *num_terms);
+int data_hamil_getnums(data_id fid, uint32_t *num_qubits, size_t *num_terms);
 
 /**
  * Get the normalization factor for the "hamil" group.
